% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_product.R
\name{dot_product}
\alias{dot_product}
\title{Calculate the dot product}
\usage{
dot_product(n1, n2)
}
\arguments{
\item{n1}{The first vector.}

\item{n2}{The second vector.}
}
\value{
Should give us a number.
}
\description{
Calculates a dot product between matrices
}
\examples{
n1 <- c(1,1,1,1,1,1)
n2 <- c(1,1,1,1,1,1)
dot_product(n1, n2)
}
