% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gramSmith.R
\name{gramSmith}
\alias{gramSmith}
\title{Gram-Smidth process in r}
\usage{
gramSmith(x1, x2)
}
\arguments{
\item{x1}{The first vector.}

\item{x2}{The second vector.}
}
\value{
Should give us a number.
}
\description{
Calculates a dot product between matrices
}
\examples{
x1 <- c(1,1,1,1,1,1)
x2 <- c(1,1,1,1,1,1)
gramSmith(x1,x2)
}
