% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_create.R
\name{matrix_create}
\alias{matrix_create}
\title{Create a Matrix}
\usage{
matrix_create(string, nr1, nr2)
}
\arguments{
\item{string}{Numbers to put in the matrix}

\item{nr1}{Number of rows}

\item{nr2}{Number of columns}
}
\value{
A matrix
}
\description{
Create a Matrix where you only need to specify which numbers you have to put in,
and how it should be organiszed.
}
\examples{
string <- c(1,1,1,1,1,1)
matrix_create(string, 2,2)
}
