% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortonormalBasis.R
\name{ortonormalBasis}
\alias{ortonormalBasis}
\title{Gives us ortonomal basis}
\usage{
ortonormalBasis(x1, x2)
}
\arguments{
\item{x1}{The first vector.}

\item{x2}{The second vector.}
}
\value{
Should give us a number.
}
\description{
Calculates a dot product between matrices
}
